+++ src/main/java/com/alaska/test/dto/LineDto.java
@@ -10,8 +10,8 @@ public class LineDto {
 	private String local_name;
 	private String data_name;
 	private String line_color;
-	private int line_width;
-	private int opacity;
+	private float line_width;
+	private float opacity;
 	private String layer_type;
 
 }
