+++ src/main/java/com/alaska/test/dto/PolygonDto.java
@@ -10,9 +10,9 @@ public class PolygonDto {
 	private String local_name;
 	private String data_name;
 	private String line_color;
-	private int line_width;
+	private float line_width;
 	private String fill_color;
-	private int opacity;
+	private float opacity;
 	private String layer_type;
 
 }
